Split lib into wasm and native
