+++ .github/workflows/ci.yaml
@@ -79,6 +79,7 @@ jobs:
           toolchain: nightly-2022-11-01
           override: true
           components: clippy
+          target: wasm32-unknown-unknown
       - uses: actions/cache@v3
         env:
           cache-name: cache-cargo
@@ -92,6 +93,7 @@ jobs:
             ${{ runner.os }}-build-
             ${{ runner.os }}-
       - run: cargo clippy --workspace --all-targets --all-features -- -D warnings -W clippy::pedantic
+      - run: cargo clippy --workspace --target wasm32-unknown-unknown --all-features -- -D warnings -W clippy::pedantic
 
   cargo_test:
     name: "cargo test"
@@ -118,6 +120,8 @@ jobs:
       - run: pip install black[d]==22.12.0
       - run: cargo test --all
       - run: cargo test --package ruff --test black_compatibility_test -- --ignored
+      - run: curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
+      - run: wasm-pack test --node
 
   maturin_build:
     name: "maturin build"
