+++ src/lib.rs
@@ -11,19 +11,10 @@
     clippy::too_many_lines
 )]
 
-use std::path::Path;
-
-use anyhow::Result;
-use path_absolutize::path_dedot;
-use rustpython_helpers::tokenize;
-use rustpython_parser::lexer::LexResult;
-use settings::{pyproject, Settings};
+use cfg_if::cfg_if;
 
 use crate::checks::Check;
-use crate::linter::check_path;
-use crate::resolver::Relativity;
-use crate::settings::configuration::Configuration;
-use crate::settings::flags;
+use crate::settings::Settings;
 use crate::source_code_locator::SourceCodeLocator;
 
 mod ast;
@@ -34,7 +25,6 @@ pub mod checks;
 pub mod checks_gen;
 pub mod cli;
 pub mod code_gen;
-pub mod commands;
 mod cst;
 mod directives;
 mod docstrings;
@@ -66,7 +56,6 @@ pub mod logging;
 pub mod mccabe;
 pub mod message;
 mod noqa;
-mod packages;
 mod pandas_vet;
 pub mod pep8_naming;
 pub mod printer;
@@ -82,55 +71,20 @@ mod ruff;
 mod rustpython_helpers;
 pub mod settings;
 pub mod source_code_locator;
-#[cfg(feature = "update-informer")]
-pub mod updates;
 mod vendored;
 pub mod visibility;
 
-/// Load the relevant `Settings` for a given `Path`.
-fn resolve(path: &Path) -> Result<Settings> {
-    if let Some(pyproject) = pyproject::find_pyproject_toml(path)? {
-        // First priority: `pyproject.toml` in the current `Path`.
-        resolver::resolve_settings(&pyproject, &Relativity::Parent, None)
-    } else if let Some(pyproject) = pyproject::find_user_pyproject_toml() {
-        // Second priority: user-specific `pyproject.toml`.
-        resolver::resolve_settings(&pyproject, &Relativity::Cwd, None)
+cfg_if! {
+    if #[cfg(not(target_family = "wasm"))] {
+        pub mod commands;
+        mod packages;
+        #[cfg(all(feature = "update-informer"))]
+        pub mod updates;
+
+        mod lib_native;
+        pub use lib_native::check;
     } else {
-        // Fallback: default settings.
-        Settings::from_configuration(Configuration::default(), &path_dedot::CWD)
+        mod lib_wasm;
+        pub use lib_wasm::check;
     }
-}
-
-/// Run Ruff over Python source code directly.
-pub fn check(path: &Path, contents: &str, autofix: bool) -> Result<Vec<Check>> {
-    // Load the relevant `Settings` for the given `Path`.
-    let settings = resolve(path)?;
-
-    // Tokenize once.
-    let tokens: Vec<LexResult> = tokenize(contents);
-
-    // Initialize the SourceCodeLocator (which computes offsets lazily).
-    let locator = SourceCodeLocator::new(contents);
-
-    // Extract the `# noqa` and `# isort: skip` directives from the source.
-    let directives = directives::extract_directives(
-        &tokens,
-        &locator,
-        directives::Flags::from_settings(&settings),
-    );
-
-    // Generate checks.
-    let checks = check_path(
-        path,
-        packages::detect_package_root(path),
-        contents,
-        tokens,
-        &locator,
-        &directives,
-        &settings,
-        autofix.into(),
-        flags::Noqa::Enabled,
-    )?;
-
-    Ok(checks)
 }
